backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data processing
pandas==2.2.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Template engine
jinja2==3.1.2

# OCR and Image Processing
pytesseract==0.3.13
pillow==11.2.1

# HTTP requests
requests==2.31.0

# AI/ML Integration
openai==1.3.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Health checks
healthcheck==1.3.0
